requirements.txt
cryptography>=3.4
pyOpenSSL>=20.0
pyfiglet>=0.8
progress>=1.5
PyQt6